merge

+++ src/routes/api/v1/media/delete/+server.ts
@@ -2,6 +2,7 @@ import { z } from 'zod';
 import { R2_BUCKET } from '$env/static/private';
 import { MediaDeleteSchema } from '$ts/common/schema';
 import s3 from '$ts/server/s3';
+import { DeleteObjectCommand } from '@aws-sdk/client-s3';
 
 export const POST = async ({ request, locals }) => {
 	// Check if the user is logged in
@@ -46,12 +47,21 @@ export const POST = async ({ request, locals }) => {
 	const key = media.url.replace(/^https:\/\/pub-3aabe8e9655b4a5eb94c0efbaa7142a1\.r2\.dev\//, '');
 
 	// Delete the media from R2 storage
-	const deleteResponse = await s3
-		.deleteObject({
-			Bucket: R2_BUCKET,
-			Key: key
-		})
-		.promise();
+	const deleteObjectCommand = new DeleteObjectCommand({
+		Bucket: R2_BUCKET,
+		Key: key
+	});
+
+	const deleteObjectResponse = await s3.send(deleteObjectCommand);
+
+	// Check if the media was successfully deleted
+	if (deleteObjectResponse.$metadata.httpStatusCode !== 204)
+		return new Response(
+			JSON.stringify({ error: 'An error occured when deleting the media from storage.' }),
+			{
+				status: 500
+			}
+		);
 
 	// Delete the media from the database
 	await locals.prisma.userMedia.delete({
