+++ src/routes/app/dashboard/profile/+page.server.ts
@@ -1,6 +1,7 @@
 import type { TilePage } from '$ts/common/types.js';
 import { R2_BUCKET } from '$env/static/private';
 import s3 from '$ts/server/s3';
+import { DeleteObjectCommand } from '@aws-sdk/client-s3';
 
 export const load = async ({ locals }) => {
 	if (!locals.user) return { files: [] };
@@ -16,16 +17,15 @@ export const load = async ({ locals }) => {
 		// Extract the key from the media URL
 		const key = url.replace(/^https:\/\/pub-3aabe8e9655b4a5eb94c0efbaa7142a1\.r2\.dev\//, '');
 
-		// Delete the media from R2 storage
-		const deleteResponse = await s3
-			.deleteObject({
-				Bucket: R2_BUCKET,
-				Key: key
-			})
-			.promise();
+		const deleteObjectCommand = new DeleteObjectCommand({
+			Bucket: R2_BUCKET,
+			Key: key
+		});
 
-		// Check if the deletion was successful
-		// if (deleteResponse.status !== 204) return false;
+		const deleteObjectResponse = await s3.send(deleteObjectCommand);
+
+		// Check if the media was successfully deleted
+		if (deleteObjectResponse.$metadata.httpStatusCode !== 204) return false;
 
 		// Delete the media from the database
 		await locals.prisma.userMedia.delete({
@@ -60,8 +60,8 @@ export const load = async ({ locals }) => {
 
 	// Get all the image URLs from the user's projects
 	const imageUrls: string[] = userProjects.flatMap((project) =>
-		project.pages.flatMap((page) =>
-			(page as TilePage)?.data?.tiles.flatMap((tile) => tile?.image ?? [])
+		project.pages.flatMap(
+			(page) => (page as TilePage)?.data?.tiles.flatMap((tile) => tile?.image ?? [])
 		)
 	);
